backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==23.0.0
httpx==0.27.2
pydantic==2.9.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pykiteconnect==5.0.1
pytest==8.3.3
